+++ src/pages/operator/index.ts
@@ -3,4 +3,4 @@ export { DeployPaymaster } from './DeployPaymaster';
 export { ConfigurePaymaster } from './ConfigurePaymaster';
 export { StakeEntryPoint } from './StakeEntryPoint';
 export { RegisterToRegistry } from './RegisterToRegistry';
-export { ManagePaymaster } from './ManagePaymaster';
+export { default as ManagePaymaster } from './ManagePaymaster';